backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.123.9
uvicorn==0.38.0
python-multipart==0.0.20
python-dotenv==1.2.1
pydantic==2.12.5
pydantic-settings==2.12.0

# Database
sqlalchemy==2.0.36
alembic==1.14.0
aiosqlite==0.20.0

# HTTP Client
httpx==0.28.1

# AI/LLM Services
groq==0.4.0

# PDF Processing
PyMuPDF==1.26.6
Pillow==12.0.0

# OCR (Optional - uncomment if needed)
# paddleocr==2.10.0
# paddlepaddle==3.0.0

# YOLO (Optional - uncomment if needed)
# ultralytics==8.3.0

# Utilities
python-dateutil==2.9.0

# FastAPI extras
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
click==8.3.1
email-validator==2.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
starlette==0.50.0
typing_extensions==4.15.0
uvloop==0.22.1
watchfiles==1.1.1

# Development
pytest==8.3.0
pytest-asyncio==0.24.0
